split gear_shop_menu into 2 TUs
I picked "render" because it seems to be doing a lot of that?
