+++ config/gear_shop_menu.yaml
@@ -56,7 +56,9 @@ segments:
     subalign: 4
     subsegments:
       - [0x0, .rodata, main/main]
+      - [0x38, .rodata, main/render]
       - [0x11c, c, main/main]
+      - [0x91d0, c, main/render]
       # Add subsegments here to capture code!
       # Syntax:
       # - [ offset_hex, c, path/file ]
